feat: installs the first test for the mock interview page
