+++ package.json
@@ -23,18 +23,23 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
+    "@types/jest": "^29.5.14",
     "@types/json-schema": "^7.0.15",
     "@types/node": "^22.15.17",
     "@types/nodemon": "^1.19.6",
     "@types/pg": "^8.15.2",
-    "@types/react": "19.1.3",
+    "@types/react": "^19.1.3",
     "@types/react-dom": "^19.1.5",
     "eslint": "^9.26.0",
     "eslint-config-next": "^15.3.2",
     "eslint-plugin-react": "^7.37.5",
     "globals": "^16.1.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
-    "ts-jest": "^29.3.3",
+    "jest-environment-jsdom": "^29.7.0",
+    "prisma": "^6.8.1",
+    "ts-jest": "^29.3.4",
+    "ts-node": "^10.9.2",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.32.1"
