+++ src/tests/dashboard/mockInterview.test.tsx
@@ -1,0 +1,16 @@
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom';
+import MockInterview from '@/app/dashboard/mockInterview/page';
+import React from "react";
+describe('MockInterviewPage', () => {
+  it('renders the mock interview questions', () => {
+    render(<MockInterview />);
+    expect(screen.getByText(/Mock Interview Practice/i)).toBeInTheDocument();
+    expect(screen.getByText(/question:/i)).toBeInTheDocument();
+  });
+
+  it('allows the user to submit a text', () => {
+    render(<MockInterview />);
+    expect(screen.)
+  })
+});
