+++ tsconfig.jest.json
@@ -1,0 +1,8 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+      "moduleResolution": "node",
+      "jsx": "react-jsx" // Required for testing JSX
+    }
+  }
+
